8.1 Play Pause Functionallity

+++ controllers/videoController.js
@@ -1,5 +1,6 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import { Collection } from "mongoose";
 
 export const home = async (req, res) => {
   try {
@@ -32,20 +33,26 @@ export const search = async (req, res) => {
 // export const videos = (req, res) =>
 //   res.render("videos", { pageTitle: "Videos" });
 
-export const getUpload = (req, res) =>
+export const getUpload = (req, res) => {
   res.render("upload", { pageTitle: "Upload" });
+};
 
 export const postUpload = async (req, res) => {
   const {
     body: { title, description },
     file: { path },
+    user: { id },
   } = req;
+
   const newVideo = await Video.create({
     fileUrl: path,
     title,
     description,
+    creator: id,
   });
-  console.log(newVideo);
+
+  req.user.videos.push(newVideo.id);
+  req.user.save();
 
   res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -55,7 +62,8 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
+
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -70,7 +78,12 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -96,7 +109,13 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findByIdAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
